requirements.txt
Django==4.2.1
django-cors-headers==4.3.1
django-heroku==0.3.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
pillow==10.2.0
psycopg2==2.9.9
